requirements.txt
matplotlib
pandas
numpy
roboticstoolbox-python